rename sha1 libraries to standard
